# docs/openapi.yaml
basePath: /api/v1
definitions:
  apierror.Response:
    properties:
      code:
        type: string
        description: Stable machine-readable error code. Clients should branch on this field rather than
          message text.
      details:
        additionalProperties:
          type: string
        type: object
        description: Optional field-level validation details keyed by request field.
      message:
        type: string
        description: Safe human-readable error summary.
      request_id:
        type: string
        description: Server-generated ID for tracing this request; also returned in X-Request-ID.
    type: object
    description: Response representation used by this API.
    example:
      code: INVALID_REQUEST
      message: Invalid request
      request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      details:
        total: must be a decimal string
...
  model.BillHeadDto:
    properties:
      id:
        type: string
        description: Unique bill identifier.
      lines:
        items:
          $ref: '#/definitions/model.BillLineDto'
        type: array
        description: Selections and server-calculated rates attached to this bill.
      payout:
        type: string
        description: Settlement payout as an exact-decimal string; absent until settled.
      settled_at:
        type: string
        description: Settlement timestamp in RFC 3339 format when settled.
      status:
        type: string
        description: Bill lifecycle state.
      total:
        example: '100.00'
        type: string
        description: Total stake as a JSON string with two fractional digits.
      user_id:
        type: string
        description: Identifier of the associated resource.
      voided_at:
        type: string
        description: Void timestamp in RFC 3339 format when voided.
    type: object
    description: BillHeadDto representation used by this API.
  model.BillLineDto:
    properties:
      betting_on:
        type: string
        description: Selected side for the associated match.
      bill_id:
        type: string
        description: Identifier of the associated resource.
      match:
        $ref: '#/definitions/model.MatchDto'
      match_id:
        type: string
        description: Identifier of the selected match.
      rate:
        example: '2.000000'
        type: string
        description: Server-calculated rate as a JSON string with six fractional digits.
    type: object
    description: BillLineDto representation used by this API.
...
  model.CreateBillLineRequest:
    properties:
      betting_on:
        type: string
        description: 'Selection: team_a or team_b.'
      match_id:
        type: string
        description: Identifier of an existing match accepting bets.
    required:
    - betting_on
    - match_id
    type: object
    description: CreateBillLineRequest representation used by this API.
  model.CreateBillRequest:
    properties:
      lines:
        items:
          $ref: '#/definitions/model.CreateBillLineRequest'
        minItems: 1
        type: array
        description: One or more match selections to include in the bill.
      total:
        example: '100.00'
        type: string
        description: Positive total stake as a decimal string; numeric JSON values are rejected.
    required:
    - lines
    - total
    type: object
    description: CreateBillRequest representation used by this API.
...
  model.MatchDto:
    properties:
      end_time:
        type: string
        description: Match end timestamp in RFC 3339 format.
      id:
        type: string
        description: Unique match identifier.
      is_draw:
        type: boolean
        description: True when the terminal result is a draw.
      start_time:
        type: string
        description: Match start timestamp in RFC 3339 format.
      team_a:
        type: string
      team_a_rate:
        example: '1.750000'
        type: string
        description: Current server-calculated rate for team_a, represented with six fractional digits.
      team_a_score:
        type: integer
      team_b:
        type: string
      team_b_rate:
        example: '2.500000'
        type: string
        description: Current server-calculated rate for team_b, represented with six fractional digits.
      team_b_score:
        type: integer
      type:
        type: string
      winner:
        type: string
        description: Winning team ID, or empty when no winner is set.
    type: object
    description: MatchDto representation used by this API.
...
host: localhost:8080
info:
  title: Intania 888 API
  version: 1.0.0-breaking
  description: "Browser and application API for Intania 888. All application routes use the /api/v1 base\
    \ path. Browser authentication uses an HttpOnly session cookie; call GET /auth/me after login to retrieve\
    \ the current profile and session-bound csrf_token. Send that token in X-CSRF-Token with browser mutations\
    \ while a session is active and send an Origin allowed by the server. Mutations are rejected when\
    \ either check fails. \n\nMoney values are JSON strings. Responses use exactly two fractional digits;\
    \ requests accept up to two. Rate values are JSON strings with up to six fractional digits in requests\
    \ and six in responses. Numeric JSON tokens are not accepted for these decimal values. \n\nAPI failures\
    \ use the shared apierror.Response shape: code, message, request_id, and optional field-level details.\
    \ The same request ID is returned in the X-Request-ID header. Dates and timestamps use the formats\
    \ documented by each field. The separate /external/* backend integration is described in docs/README.md\
    \ and is intentionally outside this browser-focused reference.\n\nThe API is rate limited to 200 requests\
    \ per minute per client IP. Exceeding the limit returns 429 TOO_MANY_REQUESTS."
paths:
...
  /bills:
    get:
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BillHeadDto'
            type: array
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'The browser session is missing or expired. Failure body follows apierror.Response:
            stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Get the authenticated user's bills
      tags:
      - Bill
      description: Returns the authenticated account’s bills in the server’s existing order. Each bill
        includes its lines, status, exact-decimal total, and any settlement or void details.
    post:
      consumes:
      - application/json
      description: Places a bill for the authenticated account. Supply a positive total and at least one
        line with a match ID and selection. The server validates the match and selection and calculates
        the applicable rates; clients must not submit rates.
      parameters:
      - description: Bill stake and selections
        in: body
        name: bill
        required: true
        schema:
          $ref: '#/definitions/model.CreateBillRequest'
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        required: true
        type: string
      produces:
      - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/model.BillHeadDto'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              id: e2a37c0d-3a7d-4ad1-a5f7-b79bf78fd6d0
              total: '100.00'
              status: pending
              lines:
              - match_id: 1d93849c-a61c-4eb1-bcf4-4d27a44002bb
                betting_on: team_a
                rate: '1.750000'
        '400':
          description: 'Bad Request Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'Forbidden Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Not Found Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '409':
          description: 'Conflict Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '422':
          description: 'Unprocessable Entity Failure body follows apierror.Response; code is stable, message
            is safe to display, and request_id matches the X-Request-ID response header. Failure body
            follows apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Place an authoritative bill
      tags:
      - Bill
  /bills/{id}:
    get:
      parameters:
      - description: Bill ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/model.BillHeadDto'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Not Found Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Get a bill by ID
      tags:
      - Bill
      description: Returns one bill only when it belongs to the authenticated account. The response includes
        line selections, server-calculated rates, settlement status, and exact-decimal amounts.
...
schemes:
- http
securityDefinitions:
  CookieSession:
    description: Browser session cookie. The cookie is HttpOnly and is normally sent by the browser automatically.
      Production uses the __Host-session cookie name. Swagger 2.0 represents this as a Cookie header.
      For browser mutations, also send X-CSRF-Token from GET /auth/me and an allowed Origin.
    in: header
    name: Cookie
    type: apiKey
swagger: '2.0'
consumes:
- application/json
produces:
- application/json
...